Restructure files for collapsible
