--- draft-packages/collapsible/KaizenDraft/Collapsible/index.ts
@@ -1,3 +1,0 @@
-export * from "./Collapsible"
-export { CollapsibleGroup } from "./CollapsibleGroup"
-export { ExpertAdviceCollapsible } from "./ExpertAdviceCollapsible"
