+++ draft-packages/collapsible/KaizenDraft/CollapsibleGroup/CollapsibleGroup.module.scss
