+++ draft-packages/collapsible/KaizenDraft/index.ts
@@ -1,0 +1,3 @@
+export * from "./Collapsible/Collapsible"
+export { CollapsibleGroup } from "./CollapsibleGroup/CollapsibleGroup"
+export { ExpertAdviceCollapsible } from "./ExpertAdviceCollapsible/ExpertAdviceCollapsible"
